requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.2
django-allauth==0.54.0
django-cors-headers==4.1.0
django-filter==23.2
django-restframework==0.0.1
django-shortcuts==1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
importlib-metadata==6.7.0
Markdown==3.4.3
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
zipp==3.15.0
